lambda-requirements.txt
astral==1.10.1
astroid==2.2.2
beautifulsoup4==4.7.1
certifi==2018.11.29
chardet==3.0.4
Click==7.0
decorator==4.3.2
docutils==0.14
future==0.17.1
geocoder==1.38.1
idna==2.7
isort==4.3.10
jmespath==0.9.3
lazy-object-proxy==1.3.1
mccabe==0.6.1
pylint==2.3.1
python-dateutil==2.8.0
ratelim==0.1.6
requests==2.20.0
six==1.12.0
soupsieve==1.8
typed-ast==1.2.0
urllib3==1.24.2
wrapt==1.11.1